+++ app/controllers/tenants/base_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Tenants
+  class BaseController < ApplicationController
+    before_action :require_tenant
+  end
+end
