+++ app/controllers/tenants/profiles_controller.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module Tenants
+  class ProfilesController < BaseController
+    include Authorization
+
+    before_action :set_membership
+    before_action -> { authorize_action!(@membership) }
+
+    def edit; end
+
+    def update
+      if @membership.update(membership_params)
+        redirect_to edit_tenant_profile_path(tenant_slug: @membership.tenant.identifier),
+                    notice: t('flash.actions.update.notice', resource: Tenant.model_name.human)
+      else
+        render :edit, status: :unprocessable_entity
+      end
+    end
+
+    private
+
+    def set_membership
+      return unless current_user
+
+      @membership = current_user.tenant_memberships.find_by(tenant: current_tenant)
+    end
+
+    def membership_params
+      params.expect(tenant_membership: [:display_name])
+    end
+  end
+end
