+++ app/controllers/tenants/users_controller.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Tenants
+  class UsersController < BaseController
+    before_action :set_user
+
+    def show
+      @membership = current_tenant.tenant_memberships.find_by!(user: @user)
+    end
+
+    private
+
+    def set_user
+      @user = User.find(params[:id])
+    end
+  end
+end
