--- app/policies/tenant_profiles_policy.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-class TenantProfilesPolicy < ApplicationPolicy
-  def edit?
-    membership_owner?
-  end
-
-  def update?
-    membership_owner?
-  end
-end
