+++ app/policies/tenants/profiles_policy.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Tenants
+  class ProfilesPolicy < ApplicationPolicy
+    def edit?
+      membership_owner?
+    end
+
+    def update?
+      membership_owner?
+    end
+  end
+end
