+++ config/routes.rb
@@ -52,7 +52,11 @@ Rails.application.routes.draw do
   scope '/:tenant_slug', as: :tenant do
     get '/', to: 'tenants#show', as: :root
 
-    get 'profile', to: 'tenant_profiles#edit', as: :users_profile
-    patch 'profile', to: 'tenant_profiles#update'
+    scope module: 'tenants' do
+      resource :profile,
+               only: %i[edit update],
+               controller: 'profiles',
+               path: 'profile'
+    end
   end
 end
